Refactor clock types.
