+++ src/Avalonia.Animation/IClock.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Avalonia.Animation
+{
+    public interface IClock : IObservable<TimeSpan>
+    {
+        bool HasSubscriptions { get; }
+        TimeSpan CurrentTime { get; }
+        PlayState PlayState { get; set; }
+    }
+}
